port some tweaks from test_w_no_lut branch to this branch

+++ src/steamcompmgr.hpp
@@ -143,7 +143,7 @@ extern float focusedWindowOffsetY;
 extern bool g_bFSRActive;
 
 extern uint32_t inputCounter;
-extern uint64_t g_lastWinSeq;
+extern uint32_t g_lastWinSeq;
 
 void nudge_steamcompmgr( void );
 void force_repaint( void );
